moved files to new cache package (#4378)
