+++ conans/test/functional/configuration/default_profile_test.py
@@ -2,7 +2,7 @@ import os
 import unittest
 
 from conans.client import tools
-from conans.client.cache import PROFILES_FOLDER
+from conans.client.cache.cache import PROFILES_FOLDER
 from conans.paths import CONANFILE
 from conans.test.utils.test_files import temp_folder
 from conans.test.utils.tools import NO_SETTINGS_PACKAGE_ID, TestClient
